fix: イベント削除ボタンをedit-eventからgroupへ移動 - 削除ボタンのみをクライアントコンポーネント化 - 削除ロジックをgroupのaction.tsに移動
